refactor: fix style in HelpPopUp for readibility(TNA-91)
